ensure annotation coordinates are floats on add and save
Normalize bounds/points x and y to float in
saveAnnotationsToFile so integer-valued values loaded from
disk round-trip as floats. Also cast UI-added point coords
explicitly to float.

+++ src/ui/imageCanvas.cpp
@@ -518,8 +518,8 @@ namespace shoecomp
                 {
                     jt::Json pt;
                     pt.setObject();
-                    pt["x"] = ic.x;
-                    pt["y"] = ic.y;
+                    pt["x"] = (float)ic.x;
+                    pt["y"] = (float)ic.y;
                     if (mode == AnnotationMode::AddPoint)
                     {
                         pt["type"] =
